chore: Move phases to prevent circular import
ConversionPhases is now moved outside of utils module so that it doesn't
cause a circular import with logger module
