+++ convert2rhel/main.py
@@ -24,8 +24,8 @@ from convert2rhel import actions, applock, backup, breadcrumbs, cli, exceptions
 from convert2rhel import logger as logger_module
 from convert2rhel import pkghandler, pkgmanager, subscription, systeminfo, utils
 from convert2rhel.actions import level_for_raw_action_data, report
+from convert2rhel.phase import ConversionPhases
 from convert2rhel.toolopts import tool_opts
-from convert2rhel.utils.phase import ConversionPhases
 
 
 loggerinst = logger_module.root_logger.getChild(__name__)
